don't mix messages and schemas
